дадання нових класів
