+++ REF1/Program.cs
@@ -1,66 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using REF1;
 
-class User
-{
-    public string Name { get; set; }
-    public string Email { get; set; }
-    public string Password { get; set; }
-
-    public void SaveToDatabase()
-    {
-        Console.WriteLine("Saving user to database...");
-    }
-
-    public void SendEmail(string message)
-    {
-        Console.WriteLine($"Sending email to {Email}: {message}");
-    }
-}
-
-class Order
-{
-    public int Id { get; set; }
-    public List<string> Products { get; set; }
-    public string PaymentMethod { get; set; }
-    public double TotalAmount { get; set; }
-
-    public void ProcessPayment()
-    {
-        if (PaymentMethod == "CreditCard")
-        {
-            Console.WriteLine("Processing credit card payment...");
-        }
-        else if (PaymentMethod == "PayPal")
-        {
-            Console.WriteLine("Processing PayPal payment...");
-        }
-        else
-        {
-            Console.WriteLine("Unsupported payment method");
-        }
-    }
-
-    public void GenerateInvoice()
-    {
-        Console.WriteLine("Generating invoice...");
-    }
-}
-
-class Inventory
-{
-    private List<string> products = new List<string> { "Laptop", "Mouse", "Keyboard" };
-
-    public void CheckStock(string product)
-    {
-        if (products.Contains(product))
-            Console.WriteLine($"{product} is in stock.");
-        else
-            Console.WriteLine($"{product} is out of stock.");
-    }
-}
-
-class Program
+public class Program
 {
     static void Main()
     {
